+++ packages/groq/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/groq
 
+## 3.0.22
+
+### Patch Changes
+
+- Updated dependencies [7168375]
+  - @ai-sdk/provider@3.0.8
+  - @ai-sdk/provider-utils@4.0.14
+
 ## 3.0.21
 
 ### Patch Changes